backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
finnhub-python==2.4.19
yfinance==0.2.36
google-genai==0.4.0
requests==2.31.0
pytz==2024.1
pydantic==2.5.3
